ООП классы и интерфейсы. Пункт 1.//ДЗ
