+++ HomeWorks/HomeWorks.sln
@@ -20,7 +20,6 @@ Global
 		{58E71BC3-337B-491C-AE54-C846581488DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{58E71BC3-337B-491C-AE54-C846581488DC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{120EF702-73C9-47A7-97FB-98F861CA765A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{120EF702-73C9-47A7-97FB-98F861CA765A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{120EF702-73C9-47A7-97FB-98F861CA765A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{120EF702-73C9-47A7-97FB-98F861CA765A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{19032AF2-3E82-4542-B42D-6F66112C6D2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
